--- src/screens/Details.tsx
@@ -1,12 +1,0 @@
-import { View, Text } from "react-native";
-import React from "react";
-
-const Details = () => {
-    return (
-        <View>
-            <Text>Details</Text>
-        </View>
-    )
-}
-
-export default Details;
